% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_Bayes.R
\name{init_W}
\alias{init_W}
\title{Generate initial values for W}
\usage{
init_W(r)
}
\arguments{
\item{r}{matrix that contains the ranked lists to be aggregated. \code{NA} indicates non-inclusion of item.}
}
\description{
Generate initial values for W.
}
